+++ relay-event-schema/src/protocol/span.rs
@@ -473,12 +473,26 @@ pub struct SpanData {
     )]
     pub gen_ai_usage_input_tokens: Annotated<Value>,
 
+    /// The input tokens used by an LLM call that were cached
+    /// (cheaper and faster than non-cached input tokens)
+    #[metastructure(field = "gen_ai.usage.input_tokens.cached")]
+    pub gen_ai_usage_input_tokens_cached: Annotated<Value>,
+
     /// The output tokens used by an LLM call (the ones the LLM actually generated)
     #[metastructure(
         field = "gen_ai.usage.output_tokens",
         legacy_alias = "ai.completion_tokens.used"
     )]
     pub gen_ai_usage_output_tokens: Annotated<Value>,
+
+    /// The output tokens used to represent the model's internal thought
+    /// process while generating a response
+    #[metastructure(field = "gen_ai.usage.output_tokens.reasoning")]
+    pub gen_ai_usage_output_tokens_reasoning: Annotated<Value>,
+
+    // Exact model used to generate the response (e.g. gpt-4o-mini-2024-07-18)
+    #[metastructure(field = "gen_ai.response.model")]
+    pub gen_ai_response_model: Annotated<Value>,
 
     /// The total cost for the tokens used
     #[metastructure(field = "gen_ai.usage.total_cost", legacy_alias = "ai.total_cost")]
@@ -1268,7 +1282,10 @@ mod tests {
             gen_ai_request_max_tokens: ~,
             gen_ai_usage_total_tokens: ~,
             gen_ai_usage_input_tokens: ~,
+            gen_ai_usage_input_tokens_cached: ~,
             gen_ai_usage_output_tokens: ~,
+            gen_ai_usage_output_tokens_reasoning: ~,
+            gen_ai_response_model: ~,
             gen_ai_usage_total_cost: ~,
             browser_name: ~,
             code_filepath: String(